refactor(trust): consolidate the hotfix into the real modules
Problem
    The trust fixes shipped as a legacy/hotfix/wrapper triplet per module:
    entroly/sufficiency.py was 19 lines that did `for name, value in
    vars(_legacy).items(): globals()[name] = value`, entroly/codec.py
    monkey-patched RecoveryReference.verify onto the legacy class at import
    time, and entroly/qccr.py re-synchronised mutable hooks back into the
    legacy module on EVERY select() call so monkeypatched bindings would be
    seen. Delivery went through 50 commits that staged base64 payload blocks
    under .github/hotfix/ and applied them from CI workflows.

Root cause
    The fixes were correct but were layered over the originals instead of
    replacing them. That is the hidden module-state drift the architecture
    rules forbid: what a caller receives depends on import order, and a
    class was mutated globally.

Change
    Folded each *_hotfix definition into its real module by AST substitution
    and deleted the shims: codec_legacy, sufficiency_legacy/_hotfix,
    qccr_legacy/_hotfix, universal_compress_legacy/_hotfix/_wrapper. One
    implementation per module, no globals() mutation, no per-call
    re-synchronisation, no monkey-patching. Removed .github/hotfix/ payload
    blocks and the three PR-scoped workflows that applied them.
    RecoveryReference.verify now checks digest AND byte_length inline.
    Cherry-picked the 5.4-5.8 content codecs, which were committed after the
    last push to PR 264 and so never reached main.

Invariant
    entroly.qccr, entroly.sufficiency, entroly.codec and
    entroly.universal_compress each have exactly one definition of every
    public name, resolvable without import-order assumptions.

Tests
    120 passed, 17 skipped across test_pr264_trust_hotfix,
    test_sufficiency_certificate, test_codec_contract, test_codecs_content,
    test_qccr, test_optimize_qccr_parity, test_json_codec_preserves_values,
    test_log_codec_collapses_repeats, test_compression_contract. ruff clean.

Measured effect
    All eight defects the hotfix targets are confirmed fixed on this branch
    and were confirmed present on main beforehand:

      json recovery == original byte stream      main: BUG -> now OK
      dict field 21+ survives schema-ization     main: BUG -> now OK
      later array record values survive          main: BUG -> now OK
      distinct error codes do not collapse       main: BUG -> now OK
      untimestamped line keeps its text          main: BUG -> now OK
      log recovery == original byte stream       main: OK  -> OK
      shell protection derived from source       main: BUG -> now OK
      verify rejects forged byte_length          main: BUG -> now OK

    Two of my own tests asserted the defective behaviour and were rewritten
    to the corrected contract rather than deleted: one required EVERY digit
    run to template to "*" (the rule that merges "code 402" with "code 500"),
    the other required SKU-0039 to be ABSENT (silent loss of a load-bearing
    identifier). Both now assert the safe rule and say why. JSON output grows
    from ~371 to ~1,909 chars on that fixture, which is the correct trade:
    the smaller output was destroying data.

Remaining limitation
    The codec registry is still not called by SDK, MCP, proxy or CLI, so the
    codecs remain a tested contract rather than shipped behaviour. Version
    surfaces still read 1.0.73 from the hotfix branch and have not been
    re-validated against the release synchroniser.
