--- .github/workflows/pr265-release-sync.yml
@@ -1,53 +1,0 @@
-name: PR265 Release Synchronization
-
-on:
-  push:
-    branches:
-      - hotfix/pr264-trust-contracts
-
-permissions:
-  contents: write
-
-concurrency:
-  group: pr265-release-sync
-  cancel-in-progress: true
-
-jobs:
-  synchronize:
-    if: github.actor != 'github-actions[bot]'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v6
-        with:
-          ref: hotfix/pr264-trust-contracts
-          fetch-depth: 0
-
-      - uses: actions/setup-python@v6
-        with:
-          python-version: '3.12'
-
-      - name: Install release-test dependencies
-        run: python -m pip install pytest packaging
-
-      - name: Synchronize approved release surfaces
-        run: python scripts/sync_release_version.py 1.0.73
-
-      - name: Verify release synchronization
-        run: |
-          python -m pytest -q \
-            tests/test_release_sync_safety.py \
-            tests/test_release_surface.py
-          git diff --check
-
-      - name: Commit generated release surfaces
-        shell: bash
-        run: |
-          if git diff --quiet && git diff --cached --quiet; then
-            echo 'Release surfaces were already synchronized.'
-            exit 0
-          fi
-          git config user.name 'entroly-release-bot'
-          git config user.email '208309368+juyterman1000@users.noreply.github.com'
-          git add -A
-          git commit -m 'chore(release): synchronize 1.0.73 surfaces'
-          git push origin HEAD:hotfix/pr264-trust-contracts
