--- entroly/universal_compress_legacy.py
@@ -1,699 +1,0 @@
-"""
-Entroly Universal Content Compressor
-=====================================
-
-Information-theoretic text compression for ANY content type — not just code.
-Works on legal documents, customer support logs, API responses, PDFs, emails,
-medical records, financial reports, and any unstructured text.
-
-Architecture:
-    Code content  → AST Skeleton + Belief Pipeline (existing Rust engine)
-    Other content → Universal Compressor (this module)
-
-Compression techniques (all pure math, zero ML inference):
-    1. TF-IDF Extractive Summarization — select highest-information sentences
-    2. Structural Compression — preserve headers/tables, compress bodies
-    3. Semantic Deduplication — SimHash near-duplicate removal
-    4. Entropy-Based Selection — keep sentences with highest information density
-
-Why this beats ModernBERT:
-    - Zero latency (no neural network inference)
-    - Deterministic (same input → same output)
-    - Works fully offline (no model weights to download)
-    - Runs in ~1ms vs ~100ms for BERT inference
-    - Scales linearly O(N) vs quadratic attention in transformers
-"""
-
-from __future__ import annotations
-
-import math
-import re
-from collections import Counter
-from typing import Any
-
-
-# ─── Content Type Detection ──────────────────────────────────────────────
-
-_CONTENT_SIGNATURES = {
-    "json": lambda t: t.strip().startswith(("{", "[")),
-    "xml": lambda t: t.strip().startswith("<") and ">" in t[:200],
-    "markdown": lambda t: bool(re.search(r'^#{1,6}\s', t[:500], re.MULTILINE)),
-    "log": lambda t: bool(re.search(r'^\d{4}[-/]\d{2}|^\[?\d{2}:\d{2}|^(DEBUG|INFO|WARN|ERROR)', t[:500], re.MULTILINE)),
-    "csv": lambda t: t.count(",") > 10 and t.count("\n") > 3 and t.count(",") / max(t.count("\n"), 1) > 2,
-    "table": lambda t: "|" in t and t.count("|") > 6 and t.count("\n") > 2,
-    "email": lambda t: bool(re.search(r'^(From|To|Subject|Date):\s', t[:500], re.MULTILINE)),
-    "legal": lambda t: any(kw in t[:2000].lower() for kw in ["whereas", "hereby", "hereinafter", "pursuant", "notwithstanding"]),
-    "stacktrace": lambda t: bool(re.search(r'(Traceback|Exception|at\s+\S+\.\S+\(|^\s+at\s)', t[:1000], re.MULTILINE)),
-}
-
-
-def detect_content_type(text: str) -> str:
-    """Detect the content type of arbitrary text.
-
-    Returns one of: json, xml, markdown, log, csv, table, email,
-    legal, stacktrace, or 'prose' as fallback.
-    """
-    if not text or len(text) < 20:
-        return "prose"
-    for content_type, detector in _CONTENT_SIGNATURES.items():
-        try:
-            if detector(text):
-                return content_type
-        except Exception:
-            continue
-    return "prose"
-
-
-# ─── Marginal Information Gain (MIG) Extractive Summarizer ───────────────
-#
-# Novel algorithm for context compression. Published nowhere — invented here.
-#
-# Standard TF-IDF scores sentences INDEPENDENTLY: each sentence gets a
-# score, pick the top-K. This is suboptimal because it ignores redundancy
-# — two sentences about the same topic both score high, but selecting
-# both wastes budget.
-#
-# MIG uses GREEDY SUBMODULAR MAXIMIZATION (Nemhauser-Wolsey-Fisher 1978):
-# At each step, select the sentence with the highest MARGINAL information
-# gain — how much NEW information it adds relative to already-selected
-# sentences. This is equivalent to maximizing the set function:
-#
-#   f(S) = I(S; D) = Σ_{w ∈ vocab} [ P(w|S) * log(P(w|S) / P(w|D)) ]
-#
-# where S = selected sentences, D = full document.
-#
-# By the Nemhauser theorem, greedy maximization of a monotone submodular
-# function achieves a (1 - 1/e) ≈ 0.63 approximation ratio — the best
-# possible in polynomial time.
-#
-# Practical impact: MIG produces summaries with ~30% more information
-# density than TF-IDF because it eliminates redundant selections.
-
-_SENTENCE_SPLIT = re.compile(r'(?<=[.!?])\s+(?=[A-Z])|(?<=\n)\s*\n')
-_WORD_TOKENIZE = re.compile(r'\b\w{2,}\b')
-
-# Common stop words to filter
-_STOP_WORDS = frozenset({
-    "the", "be", "to", "of", "and", "in", "that", "have", "it",
-    "for", "not", "on", "with", "he", "as", "you", "do", "at",
-    "this", "but", "his", "by", "from", "they", "we", "say", "her",
-    "she", "or", "an", "will", "my", "one", "all", "would", "there",
-    "their", "what", "so", "up", "out", "if", "about", "who", "get",
-    "which", "go", "me", "when", "make", "can", "like", "time", "no",
-    "just", "him", "know", "take", "people", "into", "year", "your",
-    "good", "some", "could", "them", "see", "other", "than", "then",
-    "now", "look", "only", "come", "its", "over", "think", "also",
-    "back", "after", "use", "two", "how", "our", "work", "first",
-    "well", "way", "even", "new", "want", "because", "any", "these",
-    "give", "day", "most", "us", "is", "are", "was", "were", "been",
-    "has", "had", "did", "does", "am",
-})
-
-
-def _tokenize(text: str) -> list[str]:
-    """Tokenize text into lowercase words, filtering stop words."""
-    return [
-        w.lower() for w in _WORD_TOKENIZE.findall(text)
-        if w.lower() not in _STOP_WORDS and len(w) > 1
-    ]
-
-
-def _split_sentences(text: str) -> list[str]:
-    """Split text into sentences."""
-    sentences = _SENTENCE_SPLIT.split(text)
-    cleaned = [s.strip() for s in sentences if s.strip() and len(s.strip()) > 10]
-    if len(cleaned) > 1 or len(text) < 1000:
-        return cleaned
-
-    # Logs, generated docs, and repeated prompt material often arrive as a
-    # punctuation-light blob. Treating that as one sentence makes extraction a
-    # no-op, so split into deterministic word windows for the scorer.
-    words = text.split()
-    if len(words) < 120:
-        return cleaned
-    chunk_words = 80
-    chunks = [
-        " ".join(words[i:i + chunk_words])
-        for i in range(0, len(words), chunk_words)
-        if len(words[i:i + chunk_words]) >= 10
-    ]
-    return chunks or cleaned
-
-
-def tfidf_extractive_summarize(
-    text: str,
-    target_ratio: float = 0.3,
-    min_sentences: int = 3,
-    max_sentences: int = 100,
-) -> str:
-    """Compress text using Marginal Information Gain (MIG) summarization.
-
-    Unlike standard TF-IDF (scores sentences independently), MIG uses
-    greedy submodular maximization: at each step, select the sentence
-    that adds the MOST new information relative to already-selected
-    sentences. This eliminates redundancy and maximizes information
-    coverage within the token budget.
-
-    Provably (1-1/e)-optimal by the Nemhauser-Wolsey-Fisher theorem.
-
-    Args:
-        text: Input text to summarize
-        target_ratio: Target compression ratio (0.3 = keep 30%)
-        min_sentences: Minimum sentences to keep
-        max_sentences: Maximum sentences to keep
-
-    Returns:
-        Compressed text containing the most informative sentences.
-    """
-    sentences = _split_sentences(text)
-    if len(sentences) <= min_sentences:
-        return text
-
-    # Step 1: Compute IDF weights for each unique term across all sentences
-    doc_freq: Counter = Counter()
-    sentence_terms: list[list[str]] = []
-    for sent in sentences:
-        terms = _tokenize(sent)
-        sentence_terms.append(terms)
-        for term in set(terms):
-            doc_freq[term] += 1
-
-    n_docs = len(sentences)
-    idf: dict[str, float] = {
-        term: math.log(n_docs / max(df, 1))
-        for term, df in doc_freq.items()
-    }
-
-    # Step 2: Build per-sentence weighted term vectors (TF-IDF)
-    sentence_vectors: list[Counter] = []
-    for terms in sentence_terms:
-        tf = Counter(terms)
-        vec: Counter = Counter()
-        for term, count in tf.items():
-            vec[term] = (1 + math.log(count)) * idf.get(term, 0)
-        sentence_vectors.append(vec)
-
-    # Step 3: Greedy Submodular Maximization (MIG)
-    #
-    # f(S) = coverage of unique information in the document.
-    # At each iteration, pick the sentence j that maximizes:
-    #   Δf(j|S) = Σ_w max(0, vec_j[w] - covered[w])
-    #
-    # This is the "weighted coverage" submodular function.
-    # Greedy gives (1-1/e) approximation.
-
-    target_count = max(
-        min_sentences,
-        min(max_sentences, int(len(sentences) * target_ratio))
-    )
-
-    selected_indices: list[int] = []
-    covered: Counter = Counter()  # Running coverage of selected terms
-    available = set(range(n_docs))
-
-    # Position importance: first 3 and last 2 sentences get free boost
-    position_boost = {}
-    for i in range(n_docs):
-        if i < 3:
-            position_boost[i] = 1.3
-        elif i >= n_docs - 2:
-            position_boost[i] = 1.1
-        else:
-            position_boost[i] = 1.0
-
-    # Non-annihilation floor. Extractive summarization must never reduce a
-    # non-empty document to nothing. When the IDF model degenerates — e.g.
-    # near-identical segments, where every term has df = n_docs so
-    # idf = log(n_docs/n_docs) = 0 and *every* marginal gain is 0 — the
-    # submodular stop rule `best_gain <= 0` would otherwise break before
-    # selecting a single sentence and return "". An all-zero gain is a
-    # property of the scoring model on this input, NOT evidence the text
-    # carries no information; emitting nothing is strictly dominated by
-    # emitting the position-priority lead. So we honor the submodular stop
-    # only after the floor is met, and fall back to position priority below.
-    floor = min(min_sentences, n_docs)
-
-    for _ in range(target_count):
-        if not available:
-            break
-
-        best_idx = -1
-        best_gain = -1.0
-
-        # Deterministic scan order: iterate sorted indices so gain ties
-        # resolve to the lowest index. This makes the selection (and thus
-        # compress() output) bit-identical across runs — the determinism
-        # property the engine advertises — instead of depending on set
-        # iteration order.
-        for j in sorted(available):
-            # Marginal gain: new information j adds beyond what's covered
-            gain = 0.0
-            for term, weight in sentence_vectors[j].items():
-                marginal = max(0.0, weight - covered.get(term, 0.0))
-                gain += marginal
-
-            # Apply position boost
-            gain *= position_boost.get(j, 1.0)
-
-            if gain > best_gain:
-                best_gain = gain
-                best_idx = j
-
-        if best_idx < 0:
-            break
-
-        if best_gain <= 0.0:
-            # Submodular coverage exhausted (or degenerate IDF). Stop only
-            # once the floor is satisfied; otherwise keep the highest
-            # position-priority sentence so the output is non-empty,
-            # representative, and deterministic (tie-break on lowest index).
-            if len(selected_indices) >= floor:
-                break
-            best_idx = min(
-                available,
-                key=lambda idx: (-position_boost.get(idx, 1.0), idx),
-            )
-
-        selected_indices.append(best_idx)
-        available.discard(best_idx)
-
-        # Update coverage: merge selected sentence's terms
-        for term, weight in sentence_vectors[best_idx].items():
-            covered[term] = max(covered[term], weight)
-
-    # Step 4: Reconstruct in original order for coherence
-    selected_indices.sort()
-    selected = [sentences[i] for i in selected_indices]
-    return "\n".join(selected)
-
-
-# ─── Structural Compressors ──────────────────────────────────────────────
-
-
-def compress_markdown(text: str, target_ratio: float = 0.3) -> str:
-    """Compress markdown: keep headers + first paragraph of each section."""
-    lines = text.split("\n")
-    result = []
-    in_section_body = False
-    body_lines = 0
-    max_body_lines = 3  # Keep first 3 lines per section
-
-    for line in lines:
-        stripped = line.strip()
-        if stripped.startswith("#"):
-            result.append(line)
-            in_section_body = True
-            body_lines = 0
-        elif in_section_body and body_lines < max_body_lines:
-            result.append(line)
-            if stripped:
-                body_lines += 1
-        elif not in_section_body:
-            result.append(line)
-        # Skip remaining body lines
-
-    compressed = "\n".join(result)
-    if len(compressed) < len(text) * 0.9:
-        return compressed
-    # Fallback to TF-IDF
-    return tfidf_extractive_summarize(text, target_ratio)
-
-
-def compress_email_thread(text: str) -> str:
-    """Compress email thread: keep headers + first 3 lines of each email."""
-    lines = text.split("\n")
-    result = []
-    in_header = True
-    body_lines = 0
-
-    for line in lines:
-        stripped = line.strip()
-        is_header = bool(re.match(r'^(From|To|Cc|Bcc|Subject|Date|Reply-To):\s', stripped))
-        is_separator = stripped.startswith("---") or stripped.startswith("===") or "Original Message" in stripped
-
-        if is_header or is_separator:
-            result.append(line)
-            in_header = True
-            body_lines = 0
-        elif in_header and not stripped:
-            result.append(line)
-            in_header = False
-            body_lines = 0
-        elif body_lines < 5:
-            result.append(line)
-            if stripped:
-                body_lines += 1
-        elif body_lines == 5:
-            result.append("  [...]")
-            body_lines += 1
-
-    return "\n".join(result)
-
-
-def compress_stacktrace(text: str) -> str:
-    """Compress stack traces: deduplicate repeated frames, keep boundaries."""
-    lines = text.split("\n")
-    result = []
-    seen_frames: dict[str, int] = {}
-
-    for line in lines:
-        stripped = line.strip()
-        # Always keep: exception messages, "Caused by", error descriptions
-        if any(kw in line for kw in ["Exception", "Error", "Caused by", "Traceback", "panic"]):
-            result.append(line)
-            continue
-
-        # Frame-like lines: deduplicate by normalized content
-        is_frame = bool(re.match(r'^\s+(at\s|File\s|in\s|\d+:)', line))
-        if is_frame:
-            key = re.sub(r':\d+', ':N', stripped)[:80]  # Normalize line numbers
-            if key in seen_frames:
-                seen_frames[key] += 1
-            else:
-                seen_frames[key] = 1
-                result.append(line)
-        else:
-            result.append(line)
-
-    # Append dedup summary
-    duped = sum(1 for v in seen_frames.values() if v > 1)
-    if duped:
-        result.append(f"  [{duped} repeated frames deduplicated]")
-
-    return "\n".join(result)
-
-
-def compress_csv(text: str, max_sample_rows: int = 5) -> str:
-    """Compress CSV: keep header + N sample rows + schema summary."""
-    lines = text.strip().split("\n")
-    if len(lines) <= max_sample_rows + 1:
-        return text
-
-    header = lines[0]
-    col_count = header.count(",") + 1
-    sample = lines[1:max_sample_rows + 1]
-    remaining = len(lines) - max_sample_rows - 1
-
-    result = [header] + sample
-    result.append(f"[... {remaining} more rows, {col_count} columns]")
-    return "\n".join(result)
-
-
-def compress_xml(text: str, max_depth: int = 3) -> str:
-    """Compress XML: keep structure up to max_depth, strip deep content."""
-    lines = text.split("\n")
-    result = []
-    depth = 0
-
-    for line in lines:
-        stripped = line.strip()
-        if stripped.startswith("</"):
-            depth -= 1
-        if depth <= max_depth:
-            result.append(line)
-        elif stripped.startswith("<") and not stripped.startswith("</"):
-            if depth == max_depth + 1:
-                indent = "  " * (max_depth + 1)
-                result.append(f"{indent}<!-- ... content compressed ... -->")
-        if stripped.startswith("<") and not stripped.startswith("</") and not stripped.endswith("/>"):
-            depth += 1
-
-    return "\n".join(result)
-
-
-def compress_code(text: str, target_ratio: float = 0.3) -> str:
-    """Compress source code by preserving imports and symbol skeletons.
-
-    ``universal_compress(..., content_type="code")`` used to fall through to
-    prose summarization, which no-ops on many source files because code often
-    lacks sentence punctuation. This lightweight path is deterministic and
-    language-agnostic enough for SDK fallback use.
-    """
-    lines = text.splitlines()
-    if not lines:
-        return text
-
-    keep_patterns = (
-        r"^\s*(import|from|use)\s+",
-        r"^\s*(export\s+)?(async\s+)?function\s+",
-        r"^\s*(export\s+)?(abstract\s+)?class\s+",
-        r"^\s*(export\s+)?interface\s+",
-        r"^\s*(export\s+)?type\s+",
-        r"^\s*(export\s+)?const\s+[A-Za-z_][A-Za-z0-9_]*\s*=",
-        r"^\s*(def|async\s+def|class)\s+",
-        r"^\s*(pub\s+)?(async\s+)?fn\s+",
-        r"^\s*(pub\s+)?(struct|enum|trait)\s+",
-    )
-    compiled = [re.compile(p) for p in keep_patterns]
-
-    kept: list[str] = []
-    seen: set[str] = set()
-    for line in lines:
-        stripped = line.rstrip()
-        if not stripped:
-            continue
-        if any(p.search(stripped) for p in compiled):
-            key = stripped.strip()
-            if key not in seen:
-                kept.append(stripped)
-                seen.add(key)
-
-    if not kept:
-        # Fallback: preserve the lead within the requested ratio.
-        max_chars = max(200, int(len(text) * max(0.05, target_ratio)))
-        return text[:max_chars].rstrip() + "\n...[truncated]"
-
-    out = "\n".join(kept)
-    max_chars = max(200, int(len(text) * max(0.05, target_ratio)))
-    if len(out) > max_chars:
-        out = out[:max_chars].rstrip() + "\n...[truncated]"
-    return out
-
-
-# ─── Universal Compress Entry Point ──────────────────────────────────────
-
-
-def universal_compress(
-    content: str,
-    target_ratio: float = 0.3,
-    content_type: str | None = None,
-) -> tuple[str, str, float]:
-    """Compress ANY content type using information-theoretic methods.
-
-    This is Entroly's generalization layer — works on code, prose,
-    legal docs, emails, logs, CSVs, JSON, XML, stack traces, and more.
-
-    Args:
-        content: Raw text content to compress
-        target_ratio: Target compression ratio (0.3 = keep 30%)
-        content_type: Override auto-detection (optional)
-
-    Returns:
-        (compressed_content, detected_type, savings_ratio)
-    """
-    if not content or len(content) < 200:
-        return content, "short", 0.0
-
-    # Auto-detect content type
-    ctype = content_type or detect_content_type(content)
-    if ctype in {"text", "plain"}:
-        ctype = "prose"
-
-    # Dispatch to type-specific compressor
-    _DISPATCH = {
-        "json": lambda t, r: _compress_json_universal(t),
-        "xml": lambda t, r: compress_xml(t),
-        "markdown": lambda t, r: compress_markdown(t, r),
-        "log": lambda t, r: _compress_log_universal(t),
-        "csv": lambda t, r: compress_csv(t),
-        "table": lambda t, r: _compress_table(t),
-        "email": lambda t, r: compress_email_thread(t),
-        "legal": lambda t, r: tfidf_extractive_summarize(t, r),
-        "stacktrace": lambda t, r: compress_stacktrace(t),
-        "code": lambda t, r: compress_code(t, r),
-        "prose": lambda t, r: tfidf_extractive_summarize(t, r),
-    }
-
-    compressor = _DISPATCH.get(ctype, _DISPATCH["prose"])
-    compressed = compressor(content, target_ratio)
-
-    # Non-annihilation invariant. A structural compressor must never reduce
-    # substantial input to nothing. Empty output masquerades as "perfect"
-    # compression (savings → 1.0) and would sail past the savings check
-    # below, silently destroying content — so guard it explicitly first.
-    # tfidf_extractive_summarize is itself floored to stay non-empty.
-    if not compressed or not compressed.strip():
-        compressed = tfidf_extractive_summarize(content, target_ratio)
-
-    savings = 1.0 - len(compressed) / max(len(content), 1)
-    if savings < 0.05:
-        # Compression didn't help — fallback to TF-IDF
-        compressed = tfidf_extractive_summarize(content, target_ratio)
-        savings = 1.0 - len(compressed) / max(len(content), 1)
-
-    return compressed, ctype, max(0.0, savings)
-
-
-def _compress_json_universal(text: str) -> str:
-    """Compress JSON to schema (reuses proxy_transform logic)."""
-    import json
-    stripped = text.strip()
-    try:
-        data = json.loads(stripped)
-        schema = _json_to_schema(data, depth=0, max_depth=4)
-        result = json.dumps(schema, indent=2)
-        if len(result) < len(text) * 0.6:
-            return f"[JSON schema, {len(text)} chars → {len(result)} chars]\n{result}"
-    except (json.JSONDecodeError, ValueError):
-        pass
-    return text
-
-
-# Keys whose values are the answer, not the shape. A payload is usually sent
-# because of these: the amount that was charged, the code that came back, the
-# id to correlate on. Summarising them away leaves a document that describes
-# the response without containing it.
-_LOAD_BEARING_KEY = re.compile(
-    r"(?:^|_)(?:id|ids|uuid|guid|key|ref|reference|code|codes|status|state|"
-    r"error|errors|message|msg|reason|detail|details|amount|total|balance|"
-    r"price|cost|fee|tax|currency|sku|isbn|account|invoice|order|"
-    r"timestamp|time|date|created|updated|expires|version|hash|digest|"
-    r"signature|token|url|uri|path|email|phone)(?:$|_)",
-    re.IGNORECASE,
-)
-
-
-def _json_to_schema(
-    obj: Any, depth: int = 0, max_depth: int = 4, key: str = ""
-) -> Any:
-    """Extract schema from a JSON value, preserving values of record.
-
-    Structure is what repeats; values are what the payload was sent for. This
-    elides the first and keeps the second.
-
-    Numbers are preserved verbatim. They were previously replaced by
-    ``"<int>"`` / ``"<float>"`` unconditionally, destroying exactly the class of
-    value that must survive -- identifiers, timestamps, error codes, financial
-    amounts. Measured on a payment-error fixture, ``"amount_cents": 449900``
-    became ``"amount_cents": "<int>"``: the answer to "how much was declined"
-    replaced by a type name.
-
-    That trade never paid for itself. ``"<int>"`` serialises to seven
-    characters; the value it replaced was six. The placeholder is *longer* than
-    the number for anything under seven digits, so the substitution cost tokens
-    AND lost the data. The real saving comes from array elision below, where
-    one exemplar stands in for N records.
-
-    Long strings under a load-bearing key are kept for the same reason: an
-    error message or a signed URL is the payload's point, and ``<str:87>``
-    answers nothing.
-    """
-    if depth > max_depth:
-        return "..."
-    if isinstance(obj, dict):
-        return {
-            k: _json_to_schema(v, depth + 1, max_depth, key=str(k))
-            for k, v in list(obj.items())[:20]
-        }
-    elif isinstance(obj, list):
-        if not obj:
-            return []
-        return [
-            _json_to_schema(obj[0], depth + 1, max_depth, key=key),
-            f"... ({len(obj)} items)",
-        ]
-    elif isinstance(obj, str):
-        if len(obj) <= 50 or _LOAD_BEARING_KEY.search(key or ""):
-            return obj
-        return f"<str:{len(obj)}>"
-    elif isinstance(obj, bool):
-        return obj
-    elif isinstance(obj, (int, float)):
-        return obj
-    return str(type(obj).__name__)
-
-
-# Runs that vary between repetitions of the SAME log event: counters, ids,
-# durations, addresses, hex, uuids. Collapsing them is what turns two hundred
-# distinct-looking lines back into the one event they actually are.
-_LOG_VARIABLE = re.compile(
-    r"""
-    [0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}  # uuid
-  | 0x[0-9a-fA-F]+                                                # hex literal
-  | \b[0-9a-fA-F]{16,}\b                                          # long hex/hash
-  | \b\d{1,3}(?:\.\d{1,3}){3}(?::\d+)?\b                          # ipv4[:port]
-  | \b\d+(?:\.\d+)?(?:ms|s|us|ns|kb|mb|gb)\b                      # durations/sizes
-  | \b\d[\d,._]*\b                                                # any number run
-    """,
-    re.VERBOSE | re.IGNORECASE,
-)
-
-
-def _log_template(line: str) -> str:
-    """Reduce a log line to the event it is an instance of.
-
-    Deduplication keyed on the raw line cannot collapse repeats that carry a
-    counter, and that is precisely the shape the requirement names -- hundreds
-    of downstream repeats after one causal error. Measured on a 203-line
-    fixture, 200 lines reading "request failed: connection pool exhausted
-    (retry N)" were all distinct keys, so none collapsed.
-
-    Replacing variable runs with a placeholder is the standard log-templating
-    move (Drain, He et al. ICWS 2017; LogPai). It is deliberately lossy for the
-    KEY only -- the first occurrence is emitted verbatim, so the exact values of
-    the representative line survive.
-    """
-    return _LOG_VARIABLE.sub("*", line)
-
-
-def _compress_log_universal(text: str) -> str:
-    """Collapse repeated log events, keeping the first occurrence verbatim."""
-    lines = text.split("\n")
-    seen: dict[str, int] = {}
-    result = []
-    ts_strip = re.compile(r'^\S+\s+\S+\s+')
-
-    for line in lines:
-        stripped = line.strip()
-        if not stripped:
-            continue
-        key = _log_template(ts_strip.sub("", stripped))[:100]
-        if key in seen:
-            seen[key] += 1
-        else:
-            seen[key] = 1
-            result.append(line)
-
-    final = []
-    for line in result:
-        key = _log_template(ts_strip.sub("", line.strip()))[:100]
-        count = seen.get(key, 1)
-        if count > 1:
-            final.append(f"{line}  [×{count}]")
-        else:
-            final.append(line)
-
-    return "\n".join(final)
-
-
-def _compress_table(text: str) -> str:
-    """Compress markdown/pipe-delimited tables: keep header + sample rows."""
-    lines = text.split("\n")
-    table_lines = [ln for ln in lines if "|" in ln]
-    non_table = [ln for ln in lines if "|" not in ln]
-
-    if len(table_lines) <= 6:
-        return text
-
-    # Keep header (first 2 lines usually: header + separator) + 3 sample rows
-    header = table_lines[:2]
-    sample = table_lines[2:5]
-    remaining = len(table_lines) - 5
-
-    result = non_table[:3]  # Context before table
-    result.extend(header)
-    result.extend(sample)
-    result.append(f"| ... {remaining} more rows ... |")
-    result.extend(non_table[3:6])  # Context after table
-
-    return "\n".join(result)
